requirements.txt
alembic
broadcaster[redis]
celery
gunicorn
fastapi[all]
fastapi-cache2[redis]
fastapi-users[sqlalchemy]
flower
python-jose[cryptography]
psycopg[binary]
python-dotenv
setuptools
sqlalchemy